rag-service/requirements.txt
# # Core API framework
# fastapi==0.104.1
# uvicorn==0.24.0
# pydantic==2.5.0
# python-multipart==0.0.6

# # Web scraping
# requests==2.31.0
# beautifulsoup4==4.12.2
# urllib3==2.1.0

# # ML/AI stack
# sentence-transformers==2.2.2
# transformers==4.36.0
# torch==2.8.0+cpu
# numpy==1.24.3
# scikit-learn==1.3.2

# # Vector storage and search
# faiss-cpu==1.7.4

# # Data handling & utilities
# python-dotenv==1.0.0
# pandas==2.1.4
# tqdm==4.66.1
# click==8.1.7
# colorama==0.4.6

# # Logging
# structlog==23.2.0

# # Testing
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.24.1

fastapi
uvicorn
pydantic
python-multipart

requests
beautifulsoup4
urllib3

sentence-transformers
transformers
torch
numpy
scikit-learn

faiss-cpu

python-dotenv
pandas
tqdm
click
colorama

structlog

pytest
pytest-asyncio
